Serialized some ShootingEnemy variables
This way I don't need to set every single variable in the SetUpEnemy method. Inside that method, now, I can just set the variables that are required to be set in play time and the other variables that can be known at compilation time can be set in the editor.