Enemy projectile pool fixed
The projectile GameObjects are now loaded using addressables.

A queue is used to store the different projectile pools instead of lists.

Projectiles are now created in case it was forgotten to set the type in the initial instatiation.

Deleted previous obsoleted projectile classes and fixed execution order in "ShootingEnemy".

Minor adjustment to some malfunctionig code in the player